+++ src/Psl/IO/Stream/ReadWriteHandle.php
@@ -10,11 +10,12 @@ use Psl\IO\Internal;
 /**
  * @codeCoverageIgnore
  */
-final class StreamWriteHandle implements IO\WriteHandleInterface
+final class ReadWriteHandle implements IO\ReadWriteHandleInterface
 {
+    use IO\ReadHandleConvenienceMethodsTrait;
     use IO\WriteHandleConvenienceMethodsTrait;
 
-    private IO\WriteHandleInterface $handle;
+    private IO\ReadWriteHandleInterface $handle;
 
     /**
      * @param resource|object $stream
@@ -23,7 +24,23 @@ final class StreamWriteHandle implements IO\WriteHandleInterface
      */
     public function __construct(mixed $stream)
     {
-        $this->handle = new Internal\ResourceHandle($stream, read: false, write: true, seek: false);
+        $this->handle = new Internal\ResourceHandle($stream, read: true, write: true, seek: false);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function readImmediately(?int $max_bytes = null): string
+    {
+        return $this->handle->readImmediately($max_bytes);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function read(?int $max_bytes = null, ?int $timeout_ms = null): string
+    {
+        return $this->handle->read($max_bytes, $timeout_ms);
     }
 
     /**